requirements.txt
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.8
Deprecated==1.2.7
distro-info===0.18ubuntu0.18.04.1
fail2ban==0.10.2
Flask==0.12.2
idna==2.9
itsdangerous==0.24
Jinja2==2.10
language-selector==0.1
lxml==4.4.2
MarkupSafe==1.0
netifaces==0.10.4
pdf2image==1.11.0
pika==1.1.0
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.19
pygit2==1.0.2
PyGithub==1.46
pygobject==3.26.1
PyJWT==1.7.1
pymongo==3.10.1
pyOpenSSL==19.1.0
python-apt==1.6.5+ubuntu0.2
PyYAML==3.12
requests==2.23.0
six==1.14.0
tqdm==4.41.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
Werkzeug==0.14.1
wrapt==1.12.1
xmldiff==2.4
